modify Double to BigDecimal

+++ backtest-core/src/main/java/com/yuanzhixiang/bt/engine/Counter.kt
@@ -4,19 +4,20 @@ import com.yuanzhixiang.bt.engine.domain.Symbol
 import com.yuanzhixiang.bt.engine.domain.Position
 import com.yuanzhixiang.bt.engine.domain.PendingOrder
 import com.yuanzhixiang.bt.engine.domain.TradeRecord
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 /**
  * @author Yuan Zhixiang
  */
 interface Counter {
-    fun buy(symbol: Symbol, tradeTime: LocalDateTime, price: Double, quantity: Double)
-    fun sell(symbol: Symbol, tradeTime: LocalDateTime, price: Double, quantity: Double)
+    fun buy(symbol: Symbol, tradeTime: LocalDateTime, price: BigDecimal, quantity: BigDecimal)
+    fun sell(symbol: Symbol, tradeTime: LocalDateTime, price: BigDecimal, quantity: BigDecimal)
 
     fun cancel(orderId: Long)
     fun queryPendingOrders(symbol: Symbol): List<PendingOrder>
 
-    fun queryBalance(): Double
+    fun queryBalance(): BigDecimal
     fun queryPosition(symbol: Symbol): Position
     fun queryTradeRecords(): List<TradeRecord>
